+++ src/lib.rs
@@ -10,9 +10,9 @@ extern crate num_traits;
 extern crate rand;
 
 pub mod circom_circuit;
-pub mod io;
 pub mod plonk_util;
 pub mod proofsys_type;
 pub mod prover;
 pub mod r1cs_reader;
+pub mod reader;
 pub mod utils;
