+++ src/main.rs
@@ -13,9 +13,7 @@ use zkutil::circom_circuit::{
     proving_key_json_file, r1cs_from_bin_file, r1cs_from_json_file, verification_key_json_file, witness_from_json_file, CircomCircuit,
     R1CS,
 };
-use zkutil::io;
-use zkutil::proofsys_type::ProofSystem;
-use zkutil::prover;
+use zkutil::{proofsys_type::ProofSystem, prover, reader};
 
 /// A tool to work with SNARK circuits generated by circom
 #[derive(Clap)]
@@ -207,7 +205,7 @@ fn dump_lagrange(opts: DumpLagrangeOpts) {
     };
 
     let setup =
-        prover::SetupForProver::prepare_setup_for_prover(circuit.clone(), io::load_key_monomial_form(&opts.srs_monomial_form), None)
+        prover::SetupForProver::prepare_setup_for_prover(circuit.clone(), reader::load_key_monomial_form(&opts.srs_monomial_form), None)
             .expect("prepare err");
 
     let key_lagrange_form = setup.get_srs_lagrange_form_from_monomial_form();
@@ -230,8 +228,8 @@ fn prove(opts: ProveOpts) {
 
     let setup = prover::SetupForProver::prepare_setup_for_prover(
         circuit.clone(),
-        io::load_key_monomial_form(&opts.srs_monomial_form),
-        io::maybe_load_key_lagrange_form(opts.srs_lagrange_form),
+        reader::load_key_monomial_form(&opts.srs_monomial_form),
+        reader::maybe_load_key_lagrange_form(opts.srs_lagrange_form),
     )
     .expect("prepare err");
 
@@ -247,8 +245,8 @@ fn prove(opts: ProveOpts) {
 fn verify(opts: VerifyOpts) {
     assert!(opts.proof_system == ProofSystem::Plonk, "Deprecated");
 
-    let vk = io::load_verification_key::<Bn256>(&opts.vk);
-    let proof = io::load_proof::<Bn256>(&opts.proof);
+    let vk = reader::load_verification_key::<Bn256>(&opts.vk);
+    let proof = reader::load_proof::<Bn256>(&opts.proof);
     let correct = plonk_verify(&vk, &proof).unwrap();
     if correct {
         println!("Proof is correct");
@@ -303,12 +301,12 @@ fn export_vk(opts: ExportVkOpts) {
     };
 
     let setup =
-        prover::SetupForProver::prepare_setup_for_prover(circuit.clone(), io::load_key_monomial_form(&opts.srs_monomial_form), None)
+        prover::SetupForProver::prepare_setup_for_prover(circuit.clone(), reader::load_key_monomial_form(&opts.srs_monomial_form), None)
             .expect("prepare err");
     let vk = setup.make_verification_key().unwrap();
 
     let path = std::path::Path::new(&opts.vk);
-    assert!(!path.exists(), "path for saving verification key exists: {}", path.display());z
+    assert!(!path.exists(), "path for saving verification key exists: {}", path.display());
     let writer = std::fs::File::create(&opts.vk).unwrap();
     vk.write(writer).unwrap();
     println!("Verification key saved to: {}", opts.vk);
